Crate plc_ast

This crates represents the Abstract syntax tree (AST) It is currently only a re-export of the ast module from the root, but these should eventually move here

Modules

control_statements

literals

Structs

Array

AstFactory

CompilationUnit

Date

DateAndTime

GenericBinding

Implementation

NewLines

A datastructure that stores the location of newline characters of a string. It also offers some useful methods to determine the line-number of an offset-location.

Pou

SourceRange

SourceRangeFactory

StringValue

Time

TimeOfDay

UserTypeDeclaration

Variable

VariableBlock

Enums

AccessModifier

ArgumentProperty

AstLiteral

AstStatement

DataType

DataTypeDeclaration

DirectAccessType

HardwareAccessType

LinkageType

Operator

PolymorphismMode

PouType

TypeNature

VariableBlockType

Traits

DiagnosticInfo

Functions

flatten_expression_list

flattens expression-lists and MultipliedStatements into a vec of statements. It can also handle nested structures like 2(3(4,5))

get_enum_element_name

expects a Reference or an Assignment

get_enum_element_names

enum_elements should be the statement between then enum’s brackets ( ) e.g. x : ( this, that, etc)

pre_process

Type Definitions

AstId