Enum plc_ast::ast::AstStatement

pub enum AstStatement {

    EmptyStatement {
        location: SourceRange,
        id: AstId,
    },
    DefaultValue {
        location: SourceRange,
        id: AstId,
    },
    Literal {
        kind: AstLiteral,
        location: SourceRange,
        id: AstId,
    },
    CastStatement {
        target: Box<AstStatement>,
        type_name: String,
        location: SourceRange,
        id: AstId,
    },
    MultipliedStatement {
        multiplier: u32,
        element: Box<AstStatement>,
        location: SourceRange,
        id: AstId,
    },
    QualifiedReference {
        elements: Vec<AstStatement>,
        id: AstId,
    },
    Reference {
        name: String,
        location: SourceRange,
        id: AstId,
    },
    ArrayAccess {
        reference: Box<AstStatement>,
        access: Box<AstStatement>,
        id: AstId,
    },
    PointerAccess {
        reference: Box<AstStatement>,
        id: AstId,
    },
    DirectAccess {
        access: DirectAccessType,
        index: Box<AstStatement>,
        location: SourceRange,
        id: AstId,
    },
    HardwareAccess {
        direction: HardwareAccessType,
        access: DirectAccessType,
        address: Vec<AstStatement>,
        location: SourceRange,
        id: AstId,
    },
    BinaryExpression {
        operator: Operator,
        left: Box<AstStatement>,
        right: Box<AstStatement>,
        id: AstId,
    },
    UnaryExpression {
        operator: Operator,
        value: Box<AstStatement>,
        location: SourceRange,
        id: AstId,
    },
    ExpressionList {
        expressions: Vec<AstStatement>,
        id: AstId,
    },
    RangeStatement {
        id: AstId,
        start: Box<AstStatement>,
        end: Box<AstStatement>,
    },
    VlaRangeStatement {
        id: AstId,
    },
    Assignment {
        left: Box<AstStatement>,
        right: Box<AstStatement>,
        id: AstId,
    },
    OutputAssignment {
        left: Box<AstStatement>,
        right: Box<AstStatement>,
        id: AstId,
    },
    CallStatement {
        operator: Box<AstStatement>,
        parameters: Box<Option<AstStatement>>,
        location: SourceRange,
        id: AstId,
    },
    ControlStatement {
        kind: AstControlStatement,
        location: SourceRange,
        id: AstId,
    },
    CaseCondition {
        condition: Box<AstStatement>,
        id: AstId,
    },
    ExitStatement {
        location: SourceRange,
        id: AstId,
    },
    ContinueStatement {
        location: SourceRange,
        id: AstId,
    },
    ReturnStatement {
        location: SourceRange,
        id: AstId,
    },

}

Variants

EmptyStatement

Fields

location: SourceRange

id: AstId

DefaultValue

Fields

location: SourceRange

id: AstId

Literal

Fields

kind: AstLiteral

location: SourceRange

id: AstId

CastStatement

Fields

target: Box<AstStatement>

type_name: String

location: SourceRange

id: AstId

MultipliedStatement

Fields

multiplier: u32

element: Box<AstStatement>

location: SourceRange

id: AstId

QualifiedReference

Fields

elements: Vec<AstStatement>

id: AstId

Reference

Fields

name: String

location: SourceRange

id: AstId

ArrayAccess

Fields

reference: Box<AstStatement>

access: Box<AstStatement>

id: AstId

PointerAccess

Fields

reference: Box<AstStatement>

id: AstId

DirectAccess

Fields

access: DirectAccessType

index: Box<AstStatement>

location: SourceRange

id: AstId

HardwareAccess

Fields

direction: HardwareAccessType

access: DirectAccessType

address: Vec<AstStatement>

location: SourceRange

id: AstId

BinaryExpression

Fields

operator: Operator

left: Box<AstStatement>

right: Box<AstStatement>

id: AstId

UnaryExpression

Fields

operator: Operator

value: Box<AstStatement>

location: SourceRange

id: AstId

ExpressionList

Fields

expressions: Vec<AstStatement>

id: AstId

RangeStatement

Fields

id: AstId

start: Box<AstStatement>

end: Box<AstStatement>

VlaRangeStatement

Fields

id: AstId

Assignment

Fields

left: Box<AstStatement>

right: Box<AstStatement>

id: AstId

OutputAssignment

Fields

left: Box<AstStatement>

right: Box<AstStatement>

id: AstId

CallStatement

Fields

operator: Box<AstStatement>

parameters: Box<Option<AstStatement>>

location: SourceRange

id: AstId

ControlStatement

Fields

kind: AstControlStatement

location: SourceRange

id: AstId

CaseCondition

Fields

condition: Box<AstStatement>

id: AstId

ExitStatement

Fields

location: SourceRange

id: AstId

ContinueStatement

Fields

location: SourceRange

id: AstId

ReturnStatement

Fields

location: SourceRange

id: AstId

Implementations

impl AstStatement

pub fn get_as_list(&self) -> Vec<&AstStatement>

Returns the statement in a singleton list, or the contained statements if the statement is already a list

pub fn get_location(&self) -> SourceRange

pub fn get_id(&self) -> AstId

pub fn has_direct_access(&self) -> bool

Returns true if the current statement has a direct access.

pub fn is_cast_prefix_eligible(&self) -> bool

returns true if this AST Statement is a literal or reference that can be prefixed with a type-cast (e.g. INT#23)

pub fn is_reference(&self) -> bool

Returns true if the current statement is a reference

pub fn is_qualified_reference(&self) -> bool

pub fn is_hardware_access(&self) -> bool

pub fn is_array_access(&self) -> bool

pub fn is_pointer_access(&self) -> bool

pub fn can_be_assigned_to(&self) -> bool

pub fn new_literal(kind: AstLiteral, id: AstId, location: SourceRange) -> Self

pub fn new_integer(value: i128, id: AstId, location: SourceRange) -> Self

pub fn new_real(value: String, id: AstId, location: SourceRange) -> Self

pub fn new_string( value: impl Into<String>, is_wide: bool, id: AstId, location: SourceRange ) -> Self

pub fn is_zero(&self) -> bool

Returns true if the given token is an integer or float and zero.

pub fn is_binary_expression(&self) -> bool

pub fn is_literal_array(&self) -> bool