Crate plc_ast

This crates represents the Abstract syntax tree (AST) It is currently only a re-export of the ast module from the root, but these should eventually move here

Modules

ast

control_statements

literals

mut_visitor

This module defines the AstVisitorMut trait and its associated macros. The AstVisitorMut trait provides a set of methods for mutably traversing and visiting ASTs

provider

ser

visitor

This module defines the AstVisitor trait and its associated macros. The AstVisitor trait provides a set of methods for traversing and visiting ASTs

Macros

try_from

A try_from convenience wrapper for AstNode, passed as the ex:expr argument. Will try to return a reference to the variants inner type, specified via the t:ty parameter. Converts the try_from-Result into an Option

try_from_mut

A try_from convenience wrapper for AstNode, passed as the ex:expr argument. Will try to return a reference to the variants inner type, specified via the t:ty parameter. Converts the try_from-Result into an Option

visit_all_nodes_mut