plc_ast

Module ast

Structs

Allocation

Assignment

AstFactory

AstNode

BinaryExpression

CallStatement

CastStatement

CompilationUnit

ConfigVariable

DefaultValue

DirectAccess

EmptyStatement

GenericBinding

HardwareAccess

Identifier

Implementation

Interface

JumpStatement

Represents a conditional jump from current location to a specified label

LabelStatement

Represents a location in code that could be jumbed to

MetaData

MultipliedStatement

Pou

PropertyBlock

The property container as a whole, which contains PropertyImplementations

PropertyImplementation

RangeStatement

ReferenceExpr

UnaryExpression

UserTypeDeclaration

Variable

VariableBlock

Enums

AccessModifier

ArgumentProperty

AstStatement

AutoDerefType

DataType

DataTypeDeclaration

DeclarationKind

DirectAccessType

HardwareAccessType

LinkageType

Operator

PolymorphismMode

PouType

PropertyKind

ReferenceAccess

TypeNature

VariableBlockType

Functions

flatten_expression_list

flattens expression-lists and MultipliedStatements into a vec of statements. It can also handle nested structures like 2(3(4,5))

get_enum_element_name

expects a Reference or an Assignment

get_enum_element_names

enum_elements should be the statement between then enum’s brackets ( ) e.g. x : ( this, that, etc)

pre_process

steal_expression_list

Type Aliases

AstId