Module plc_ast::ast

Structs

Assignment

AstFactory

AstNode

BinaryExpression

CallStatement

CastStatement

CompilationUnit

DefaultValue

DirectAccess

EmptyStatement

GenericBinding

HardwareAccess

Implementation

JumpStatement

Represents a conditional jump from current location to a specified label

LabelStatement

Represents a location in code that could be jumbed to

MultipliedStatement

Pou

RangeStatement

ReferenceExpr

UnaryExpression

UserTypeDeclaration

Variable

VariableBlock

Enums

AccessModifier

ArgumentProperty

AstStatement

DataType

DataTypeDeclaration

DirectAccessType

HardwareAccessType

LinkageType

Operator

PolymorphismMode

PouType

ReferenceAccess

TypeNature

VariableBlockType

Functions

flatten_expression_list

flattens expression-lists and MultipliedStatements into a vec of statements. It can also handle nested structures like 2(3(4,5))

get_enum_element_name

expects a Reference or an Assignment

get_enum_element_names

enum_elements should be the statement between then enum’s brackets ( ) e.g. x : ( this, that, etc)

pre_process

Type Aliases

AstId