Struct plc_ast::ast::AstNode

pub struct AstNode {
    pub stmt: AstStatement,
    pub id: AstId,
    pub location: SourceLocation,
}

Fields
stmt: AstStatementid: AstIdlocation: SourceLocation
Implementations

impl AstNode

pub fn get_as_list(&self) -> Vec<&AstNode>

Returns the statement in a singleton list, or the contained statements if the statement is already a list

pub fn get_location(&self) -> SourceLocation

pub fn set_location(&mut self, location: SourceLocation)

pub fn get_id(&self) -> AstId

pub fn get_stmt(&self) -> &AstStatement

pub fn get_stmt_peeled(&self) -> &AstStatement

Similar to AstNode::get_stmt with the exception of peeling parenthesized expressions. For example if called on ((1)) this function would return a AstStatement::Literal ignoring the parenthesized expressions altogether.

pub fn has_direct_access(&self) -> bool

Returns true if the current statement has a direct access.

pub fn is_cast_prefix_eligible(&self) -> bool

returns true if this AST Statement is a literal or reference that can be prefixed with a type-cast (e.g. INT#23)

pub fn is_flat_reference(&self) -> bool

Returns true if the current statement is a flat reference (e.g. a)

pub fn get_flat_reference_name(&self) -> Option<&str>

Returns the reference-name if this is a flat reference like a, or None if this is no flat reference

pub fn get_parent_name_of_reference(&self) -> Option<&str>

pub fn get_label_name(&self) -> Option<&str>

pub fn is_empty_statement(&self) -> bool

pub fn is_assignment(&self) -> bool

pub fn is_output_assignment(&self) -> bool

pub fn is_reference(&self) -> bool

pub fn is_call(&self) -> bool

pub fn is_hardware_access(&self) -> bool

pub fn is_array_access(&self) -> bool

pub fn is_pointer_access(&self) -> bool

pub fn is_paren(&self) -> bool

pub fn is_expression_list(&self) -> bool

pub fn can_be_assigned_to(&self) -> bool

pub fn new(stmt: AstStatement, id: AstId, location: SourceLocation) -> AstNode

pub fn new_literal( kind: AstLiteral, id: AstId, location: SourceLocation, ) -> AstNode

pub fn new_integer(value: i128, id: AstId, location: SourceLocation) -> AstNode

pub fn new_real(value: String, id: AstId, location: SourceLocation) -> AstNode

pub fn new_string( value: impl Into<String>, is_wide: bool, id: AstId, location: SourceLocation, ) -> AstNode

pub fn is_zero(&self) -> bool

Returns true if the given token is an integer or float and zero.

pub fn is_binary_expression(&self) -> bool

pub fn is_literal_array(&self) -> bool

pub fn is_literal(&self) -> bool

pub fn is_literal_integer(&self) -> bool

pub fn get_literal_integer_value(&self) -> Option<i128>

pub fn is_identifier(&self) -> bool

pub fn is_default_value(&self) -> bool

pub fn negate(self: AstNode, id_provider: IdProvider) -> AstNode

Negates the given element by adding it to a not expression

pub fn is_template(&self) -> bool